feat: navigation setup
