+++ package.json
@@ -11,8 +11,12 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.3",
+    "@react-navigation/native-stack": "^6.9.9",
     "react": "18.2.0",
-    "react-native": "0.71.2"
+    "react-native": "0.71.2",
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-screens": "^3.19.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
