+++ src/navigation/types.ts
@@ -1,0 +1,4 @@
+export type RootStackParamList = {
+  Weather: undefined;
+  Details: {city: Record<string, unknown>}; // @TODO: complete city type
+};
