+++ src/screens/Details.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import {SafeAreaView, StatusBar, useColorScheme} from 'react-native';
+import {Colors} from 'react-native/Libraries/NewAppScreen';
+import type {NativeStackScreenProps} from '@react-navigation/native-stack';
+import {RootStackParamList} from '../navigation/types';
+
+type Props = NativeStackScreenProps<RootStackParamList, 'Details'>;
+
+function DetailsScreen({route}: Props): JSX.Element {
+  const isDarkMode = useColorScheme() === 'dark';
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  const {city} = route.params;
+  // @TODO: show city details
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar
+        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
+        backgroundColor={backgroundStyle.backgroundColor}
+      />
+    </SafeAreaView>
+  );
+}
+
+export default DetailsScreen;
